Move AutoPlayer to core and improve placement search
